grequest callback return values and Fortran status
A little background first: the status object being updated
here based on the callbacks' behavior is an internal
req->req_status.  At this level we're not directly updating
the top level MPI_Status from the user.  And the design is
to put the callback return values in the req->req_status.MPI_ERROR.
(and whether that propagates back into the top level MPI_Status
is handled elsewhere, where the code knows whether it's in
a MPI_Waitall or MPI_Wait etc).  And the way the callback return
is put into those MPI_ERROR fields is that it's initialized to
SUCCESS, and it's only updated if the callback returns failure.

The problems fixed here are:

1. for Fortran the callback uses a stack variable fstatus that's
a Fortran status object, then copies those values back into the
req->req_status.  But that steps on the design described above:
it writes a garbage value from the stack into the MPI_ERROR field
when it was supposed to leave the value untouched.

If I make a test where I manually put garbage into fstatus it does
make its way up the chain and make MPI_Wait fail.  So I fixed this
by initializing fstatus with an MPI_Status_c2f

2. the free_fn() callback never had its return value placed into
any of the status.MPI_ERROR locations, and the standard specifically
calls out that it's the free_fn() not the query_fn() return value
that's supposed to go into status.MPI_ERROR.

Signed-off-by: Mark Allen <markalle@us.ibm.com>