decide CONVERTOR_CUDA flag based on first byte of buffer
There's a cuda check in the converter that checks one address and
caches a flag for CONVERTOR_CUDA.  But it was just checking
pUserBuf which isn't necessarily anywhere near the data, and
doesn't even need to be a legal address for that matter.  MPI
datatype displacements could put the actual data anywhere.

I added a lookup of the first non-loop pElem in the datatype
description to use its .disp to find the first byte of data
in the user's buf.

The code is probably being overly cautious, when it could just
be coded as
    pElem = &(convertor->use_desc->desc[0]);
    while (!(pElem->elem.common.flags & OPAL_DATATYPE_FLAG_DATA)) {
        ++pElem;
    }
    // use pElem->elem.disp
But I was reluctant to write code that would run away in that while
loop and fail ungracefully if the description array was bad for some
reason or let it select a wrong first displacement if degenerate loops
were a possibility in the description.

Signed-off-by: Mark Allen <markalle@us.ibm.com>